add ArtistController and ArtistService methods

+++ src/main/java/ch/bossgroup/podcastmanager/artist/ArtistService.java
@@ -3,6 +3,7 @@ package ch.bossgroup.podcastmanager.artist;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 public class ArtistService {
@@ -14,7 +15,7 @@ public class ArtistService {
     }
 
     public List<Artist> getAllArtists() {
-        return repository.findAll();
+        return repository.findByOrderById();
     }
 
     public Artist getArtist(Long id) {
@@ -25,20 +26,20 @@ public class ArtistService {
         return repository.save(artist);
     }
 
-    public Artist updateArtist(Artist artist, Long id) {
-        return repository.findById(id)
-                .map(originalArtist -> {
-                    originalArtist.setId(id);
-                    originalArtist.setFirstname(artist.getFirstname());
-                    originalArtist.setLastname(artist.getLastname());
-                    return repository.save(originalArtist);
-                })
-                .orElse(null);
+    public Artist editArtist(Artist artist, Long id) {
+        if (repository.existsById(id)) {
+            Artist newArtist = new Artist(id, artist.getFirstname(), artist.getLastname());
+            return repository.save(newArtist);
+        } else {
+            throw new NoSuchElementException("Object with Id: " + id + " does not exist!");
+        }
     }
 
     public void deleteArtist(Long id) {
-        repository.deleteById(id);
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+        } else {
+            throw new NoSuchElementException("Object with Id: " + id + " does not exist!");
+        }
     }
-
-
 }
